[MINOR][PYTHON][TESTS] Move a test out of parity tests
### What changes were proposed in this pull request?
Move a test out of parity tests

### Why are the changes needed?
it is not tested in Spark Classic, not a parity test

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #46914 from zhengruifeng/move_a_non_parity_test.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
